worker: add image_build.go
